+++ src/index.js
@@ -1,24 +1,22 @@
 #!/usr/bin/env node
 const Logger = require('./utils/Logger');
 const CLI = require('./utils/CLI');
-const Config = require('./utils/Config');
+const config = require('./utils/Config');
 const Template = require('./utils/Template');
 
-const { COMMANDS, TEMPLATES, TEMPLATES_ALIAS } = require('./constants');
+const { COMMANDS, TEMPLATES, TEMPLATES_ALIAS, OPTIONS_ALIAS } = require('./constants');
 
 const getReactTsComponentTemplate = require('./templates/react-ts-component');
 
 const init = () => {
-  const config = new Config();
-
   const command = CLI.getCommand();
-  const cliTemplate = CLI.getTemplate();
-  const cliPath = CLI.getPath();
-
-  const name = CLI.getComponentName(cliPath);
 
   if (command === COMMANDS.GENERATE_CUSTOM_TEMPLATE) {
-    const templates = Template.getCustomTemplates();
+    const cliTemplate = CLI.getTemplate();
+    const cliPath = CLI.getPath();
+    const name = CLI.getComponentName(cliPath);
+
+    const templates = Template.getCustomTemplates(config.get());
     const template = templates[cliTemplate];
 
     if (template) {
@@ -31,6 +29,10 @@ const init = () => {
   }
 
   if (command === COMMANDS.GENERATE_DEFAULT_TEMPLATE) {
+    const cliTemplate = CLI.getTemplate();
+    const cliPath = CLI.getPath();
+    const name = CLI.getComponentName(cliPath);
+
     switch (cliTemplate) {
       case TEMPLATES_ALIAS[TEMPLATES.REACT_TS_COMPONENTS]: {
         const options = config.getTemplateConfig(TEMPLATES.REACT_TS_COMPONENTS);
@@ -43,6 +45,22 @@ const init = () => {
       default:
         throw new Error('Unknown template');
     }
+
+    return;
+  }
+
+  if (command === COMMANDS.SET_OPTIONS) {
+    const options = CLI.getParams();
+
+    Object.entries(options).forEach(([key, value]) => {
+      const isExist = key in OPTIONS_ALIAS;
+
+      if (isExist) {
+        config.writeOption(key, value);
+      } else {
+        throw new Error('Unknown option for set');
+      }
+    });
 
     return;
   }