+++ src/utils/Config.js
@@ -4,19 +4,21 @@ const FS = require('./FS');
 class Config {
   constructor() {
     this.config = {};
+    this.configPath = path.resolve(__dirname, '../config.json');
 
     this.set();
   }
 
   set() {
     try {
-      const defaultConfigPath = path.resolve(__dirname, '../default.config.json');
-
-      // @todo get loca config and merge
-      this.config = JSON.parse(FS.readFileSync(defaultConfigPath));
+      this.config = JSON.parse(FS.readFileSync(this.configPath));
     } catch (err) {
       throw new Error(`Error set config ${err.message}`);
     }
+  }
+
+  get() {
+    return this.config;
   }
 
   getTemplateConfig(name) {
@@ -24,6 +26,21 @@ class Config {
 
     return template ?? {};
   }
+
+  writeOption(key, value) {
+    try {
+      const file = JSON.parse(FS.readFileSync(this.configPath));
+      file[key] = value;
+
+      const updatedJson = JSON.stringify(file, null, 2);
+
+      FS.createFile(this.configPath, updatedJson);
+    } catch (err) {
+      throw new Error(`Error write option ${err.message}`);
+    }
+  }
 }
 
-module.exports = Config;
+const config = new Config();
+
+module.exports = config;